# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Signal AEO Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/gbp-snippet:
    get:
      operationId: getClientGbpSnippet
      tags: [clients]
      summary: Get GBP snippet and verification status for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: GBP snippet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GbpSnippet"
  /clients/{id}/aeo-summary:
    get:
      operationId: getClientAeoSummary
      tags: [clients]
      summary: AEO keyword summary for a client with before/after ranking dates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AEO summary per client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientAeoSummary"
  /keywords:
    get:
      operationId: getKeywords
      tags: [keywords]
      summary: List keywords
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: tierLabel
          in: query
          schema:
            type: string
            enum: [aeo, seo, both]
      responses:
        "200":
          description: List of keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
    post:
      operationId: createKeyword
      tags: [keywords]
      summary: Create a keyword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeywordBody"
      responses:
        "201":
          description: Keyword created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
  /keywords/{id}:
    patch:
      operationId: updateKeyword
      tags: [keywords]
      summary: Update a keyword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKeywordBody"
      responses:
        "200":
          description: Updated keyword
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
    delete:
      operationId: deleteKeyword
      tags: [keywords]
      summary: Delete a keyword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /devices:
    get:
      operationId: getDevices
      tags: [devices]
      summary: List all Android devices in the farm
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [available, offline, in_use]
      responses:
        "200":
          description: Device list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Register a new device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeviceBody"
      responses:
        "201":
          description: Device registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}:
    patch:
      operationId: updateDevice
      tags: [devices]
      summary: Update device status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeviceBody"
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
...
  /proxies:
    get:
      operationId: getProxies
      tags: [proxies]
      summary: List proxies
      parameters:
        - name: proxyType
          in: query
          schema:
            type: string
            enum: [residential, mobile]
      responses:
        "200":
          description: Proxy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proxy"
    post:
      operationId: createProxy
      tags: [proxies]
      summary: Add a proxy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProxyBody"
      responses:
        "201":
          description: Proxy added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proxy"
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        businessName:
          type: string
        gmbUrl:
          type: string
          nullable: true
        websiteUrl:
          type: string
          nullable: true
        publishedAddress:
          type: string
          nullable: true
        searchAddress:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        planName:
          type: string
          nullable: true
        addressType:
          type: integer
          nullable: true
        placeId:
          type: string
          nullable: true
        locationRef:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, businessName, status, createdAt]
    CreateClientBody:
      type: object
      properties:
        businessName:
          type: string
        gmbUrl:
          type: string
          nullable: true
        websiteUrl:
          type: string
          nullable: true
        publishedAddress:
          type: string
          nullable: true
        searchAddress:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
          default: active
        planName:
          type: string
          nullable: true
        addressType:
          type: integer
          nullable: true
        placeId:
          type: string
          nullable: true
        locationRef:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
      required: [businessName]
    UpdateClientBody:
      type: object
      properties:
        businessName:
          type: string
        gmbUrl:
          type: string
          nullable: true
        websiteUrl:
          type: string
          nullable: true
        publishedAddress:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        planName:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
    GbpSnippet:
      type: object
      properties:
        clientId:
          type: integer
        businessName:
          type: string
        gmbUrl:
          type: string
          nullable: true
        placeId:
          type: string
          nullable: true
        verificationStatus:
          type: string
          enum: [verified, pending, failed]
        publishedAddress:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        mapsPresence:
          type: string
          nullable: true
        lastChecked:
          type: string
          format: date-time
          nullable: true
      required: [clientId, businessName, verificationStatus]
    ClientAeoSummary:
      type: object
      properties:
        clientId:
          type: integer
        businessName:
          type: string
        aeoKeywords:
          type: array
          items:
            $ref: "#/components/schemas/AeoKeywordSummary"
        totalClicksDelivered:
          type: integer
        averageRankingPosition:
          type: number
          nullable: true
        startDate:
          type: string
          format: date-time
          nullable: true
        lastSessionDate:
          type: string
          format: date-time
          nullable: true
      required: [clientId, businessName, aeoKeywords, totalClicksDelivered]
    AeoKeywordSummary:
      type: object
      properties:
        keywordId:
          type: integer
        keywordText:
          type: string
        initialRankingDate:
          type: string
          format: date-time
          nullable: true
        initialRankingPosition:
          type: integer
          nullable: true
        currentRankingPosition:
          type: integer
          nullable: true
        clicksDelivered:
          type: integer
        verificationStatus:
          type: string
      required: [keywordId, keywordText, clicksDelivered, verificationStatus]
    Keyword:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        keywordText:
          type: string
        tierLabel:
          type: string
          enum: [aeo, seo, both]
        isActive:
          type: boolean
        isPrimary:
          type: integer
        clickCount:
          type: integer
        last30DaysClickCount:
          type: integer
        backlinkCount:
          type: integer
        webType:
          type: integer
        keywordType:
          type: integer
        verificationStatus:
          type: string
          enum: [pending, verified, failed]
        avgScroll:
          type: integer
      required: [id, clientId, keywordText, tierLabel, isActive, isPrimary, clickCount, last30DaysClickCount, backlinkCount, verificationStatus]
    CreateKeywordBody:
      type: object
      properties:
        clientId:
          type: integer
        keywordText:
          type: string
        tierLabel:
          type: string
          enum: [aeo, seo, both]
          default: aeo
        isActive:
          type: boolean
          default: true
        isPrimary:
          type: integer
          default: 1
        webType:
          type: integer
          default: 1
        keywordType:
          type: integer
          default: 1
      required: [clientId, keywordText]
    UpdateKeywordBody:
      type: object
      properties:
        keywordText:
          type: string
        isActive:
          type: boolean
        verificationStatus:
          type: string
          enum: [pending, verified, failed]
        tierLabel:
          type: string
          enum: [aeo, seo, both]
...
    Device:
      type: object
      properties:
        id:
          type: integer
        deviceIdentifier:
          type: string
        model:
          type: string
        status:
          type: string
          enum: [available, offline, in_use]
        retiredToday:
          type: boolean
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        sessionsToday:
          type: integer
          nullable: true
      required: [id, deviceIdentifier, model, status, retiredToday]
    CreateDeviceBody:
      type: object
      properties:
        deviceIdentifier:
          type: string
        model:
          type: string
        status:
          type: string
          enum: [available, offline, in_use]
          default: available
      required: [deviceIdentifier, model]
    UpdateDeviceBody:
      type: object
      properties:
        status:
          type: string
          enum: [available, offline, in_use]
        retiredToday:
          type: boolean
        model:
          type: string
...
    Proxy:
      type: object
      properties:
        id:
          type: integer
        proxyUrl:
          type: string
        proxyType:
          type: string
          enum: [residential, mobile]
        sessionCount:
          type: integer
        lastUsed:
          type: string
          format: date-time
          nullable: true
      required: [id, proxyUrl, proxyType, sessionCount]
    CreateProxyBody:
      type: object
      properties:
        proxyUrl:
          type: string
        proxyType:
          type: string
          enum: [residential, mobile]
          default: residential
      required: [proxyUrl, proxyType]
...