# lib/api-zod/src/generated/types/updateClientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Signal AEO Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateClientBodyStatus } from "./updateClientBodyStatus";

export interface UpdateClientBody {
  businessName?: string;
  gmbUrl?: string | null;
  websiteUrl?: string | null;
  publishedAddress?: string | null;
  city?: string | null;
  state?: string | null;
  status?: UpdateClientBodyStatus;
  planName?: string | null;
  contactEmail?: string | null;
}